requirements.txt
streamlit
requests
pandas
python-dotenv
beautifulsoup4
lxml
python-dateutil
plotly

